+++ company/admin.py
@@ -2,16 +2,17 @@ import datetime
 
 from directory_constants.urls import domestic
 
-from django.contrib import admin
+from django.contrib import admin, messages
 from django.conf.urls import url
 from django.core.urlresolvers import reverse_lazy
 from django.core.signing import Signer
+from django.http import HttpResponse
 from django.template.response import TemplateResponse
 from django.views.generic import FormView
 from django import forms
 
 from core.helpers import generate_csv_response
-from company import models
+from company import helpers, models
 from company.forms import EnrolCompanies, UploadExpertise
 
 
@@ -83,7 +84,7 @@ class PublishByCompanyHouseNumberForm(forms.Form):
 
     options = (
         ("investment_support_directory", "Investment Support Directory"),
-        ("find_a_supplier", "Find a Supplier"),
+        ("find_a_supplier", "Find a CompanyUser"),
     )
 
     directories = forms.MultipleChoiceField(
@@ -212,3 +213,56 @@ class CollaborationInviteAdmin(admin.ModelAdmin):
     search_fields = ('uuid', 'company__name', 'company__number', 'collaborator_email')
     list_display = ('uuid', 'collaborator_email', 'company')
     list_filter = ('accepted', 'role')
+
+
+@admin.register(models.CompanyUser)
+class CompanyUserAdmin(admin.ModelAdmin):
+
+    search_fields = (
+        'sso_id', 'name', 'mobile_number', 'company_email', 'company__name',
+        'company__description', 'company__number', 'company__website',
+    )
+    readonly_fields = ('created', 'modified',)
+    actions = ['download_csv', 'resend_letter']
+
+    def download_csv(self, request, queryset):
+        """
+        Generates CSV report of all suppliers, with company details included.
+        """
+        response = HttpResponse(content_type='text/csv')
+        response['Content-Disposition'] = (
+            'attachment; filename="find-a-buyer_suppliers_{}.csv"'.format(
+                datetime.datetime.now().strftime("%Y%m%d%H%M%S")
+            )
+        )
+        helpers.generate_company_users_csv(file_object=response, queryset=queryset)
+        return response
+
+    download_csv.short_description = (
+        "Download CSV report for selected suppliers"
+    )
+
+    def resend_letter(self, request, queryset):
+        total_selected_users = queryset.count()
+        not_verified_users_queryset = queryset.select_related(
+            'company'
+        ).exclude(
+                company__verified_with_code=True
+        )
+        not_verified_users_count = not_verified_users_queryset.count()
+
+        for supplier in not_verified_users_queryset:
+            helpers.send_verification_letter(supplier.company)
+
+        messages.success(
+            request,
+            'Verification letter resent to {} users'.format(
+                not_verified_users_count
+            )
+        )
+        messages.warning(
+            request,
+            '{} users skipped'.format(
+                total_selected_users - not_verified_users_count
+            )
+        )