+++ company/apps.py
@@ -8,10 +8,6 @@ class CompanyConfig(AppConfig):
     def ready(self):
         from company import signals
         pre_save.connect(
-            receiver=signals.publish_companies_that_meet_criteria,
-            sender='company.Company'
-        )
-        pre_save.connect(
             receiver=signals.store_date_published,
             sender='company.Company'
         )
@@ -35,25 +31,9 @@ class CompanyConfig(AppConfig):
             receiver=signals.delete_company_elasticsearch_document,
             sender='company.Company',
         )
-        post_save.connect(
-            receiver=signals.send_account_ownership_transfer_notification,
-            sender='company.OwnershipInvite'
-        )
-        post_save.connect(
-            receiver=signals.send_account_collaborator_notification,
-            sender='company.CollaboratorInvite'
-        )
         pre_save.connect(
             receiver=signals.set_non_companies_house_number,
             sender='company.Company'
-        )
-        post_save.connect(
-            receiver=signals.create_collaboration_invite_from_ownership_invite,
-            sender='company.OwnershipInvite'
-        )
-        post_save.connect(
-            receiver=signals.create_collaboration_invite_from_collaborator_invite,
-            sender='company.CollaboratorInvite'
         )
         post_save.connect(
             receiver=signals.send_new_invite_collaboration_notification,