+++ company/management/commands/retrieve_missing_company_details.py
@@ -1,3 +1,4 @@
+from collections import Counter
 from datetime import datetime
 
 from directory_constants import company_types
@@ -5,30 +6,22 @@ from directory_constants import company_types
 from django.core.management.base import BaseCommand
 from django.db.models import Q
 
-from company import helpers, models
+from company import models
+from core.helpers import get_companies_house_profile
 
 
 class Command(BaseCommand):
     help = 'Retrieves missing data of companies such as date of creation'
 
     def handle(self, *args, **options):
-        queryset = models.Company.objects.filter(
-            Q(company_type=company_types.COMPANIES_HOUSE) &
-            (
-                Q(date_of_creation__isnull=True) |
-                Q(address_line_1__isnull=True) |
-                Q(address_line_1='')
-            )
-        )
-        failed = 0
-        succeded = 0
+        missing_data_query = Q(date_of_creation__isnull=True) | Q(address_line_1__isnull=True) | Q(address_line_1='')
+        queryset = models.Company.objects.filter(Q(company_type=company_types.COMPANIES_HOUSE) & missing_data_query)
+        counter = Counter()
         for company in queryset:
             try:
-                profile = helpers.get_companies_house_profile(company.number)
+                profile = get_companies_house_profile(company.number)
                 if profile.get('date_of_creation'):
-                    company.date_of_creation = datetime.strptime(
-                        profile['date_of_creation'], '%Y-%m-%d'
-                    )
+                    company.date_of_creation = datetime.strptime(profile['date_of_creation'], '%Y-%m-%d')
                 if profile.get('registered_office_address'):
                     address = profile['registered_office_address']
                     company.address_line_1 = address.get('address_line_1', '')
@@ -37,11 +30,10 @@ class Command(BaseCommand):
                     company.po_box = address.get('po_box', '')
                     company.postal_code = address.get('postal_code', '')
                 company.save()
-                message = f'Company {company.name} updated'
-                self.stdout.write(self.style.SUCCESS(message))
-                succeded += 1
-            except Exception as e:
-                self.stdout.write(self.style.ERROR(e))
-                failed += 1
-        self.stdout.write(self.style.SUCCESS(f'{succeded} companies updated'))
-        self.stdout.write(self.style.WARNING(f'{failed} companies failed'))
+                self.stdout.write(self.style.SUCCESS(f'Company {company.name} updated'))
+                counter['success'] += 1
+            except Exception as execption:
+                self.stdout.write(self.style.ERROR(execption))
+                counter['failure'] += 1
+        self.stdout.write(self.style.SUCCESS(f'{counter["success"]} companies updated'))
+        self.stdout.write(self.style.WARNING(f'{counter["failure"]} companies failed'))